+++ Utilities/gdcm/Source/Common/gdcmMD5.h
@@ -37,7 +37,7 @@ public :
   MD5();
   ~MD5();
 
-  static bool Compute(const char *buffer, unsigned long buf_len, char digest_str[33]);
+  static bool Compute(const char *buffer, size_t buf_len, char digest_str[33]);
 
   static bool ComputeFile(const char *filename, char digest_str[33]);
