+++ Utilities/gdcm/Source/InformationObjectDefinition/gdcmDefs.cxx
@@ -256,9 +256,9 @@ Type Defs::GetTypeFromTag(const File& file, const Tag& tag) const
   const IOD &iod = iods.GetIOD( iodname );
   const Macros &macros = GetMacros();
 
-  unsigned int niods = iod.GetNumberOfIODs();
+  size_t niods = iod.GetNumberOfIODs();
   // Iterate over each iod entry in order:
-  for(unsigned int idx = 0; idx < niods; ++idx)
+  for(size_t idx = 0; idx < niods; ++idx)
     {
     const IODEntry &iodentry = iod.GetIODEntry(idx);
     const char *ref = iodentry.GetRef();
@@ -292,10 +292,10 @@ bool Defs::Verify(const File& file) const
 
   //std::cout << iod << std::endl;
   //std::cout << iod.GetIODEntry(14) << std::endl;
-  unsigned int niods = iod.GetNumberOfIODs();
+  size_t niods = iod.GetNumberOfIODs();
   bool v = true;
   // Iterate over each iod entry in order:
-  for(unsigned int idx = 0; idx < niods; ++idx)
+  for(size_t idx = 0; idx < niods; ++idx)
     {
     const IODEntry &iodentry = iod.GetIODEntry(idx);
     const char *ref = iodentry.GetRef();
@@ -327,10 +327,10 @@ bool Defs::Verify(const DataSet& ds) const
 
   //std::cout << iod << std::endl;
   //std::cout << iod.GetIODEntry(14) << std::endl;
-  unsigned int niods = iod.GetNumberOfIODs();
+  size_t niods = iod.GetNumberOfIODs();
   bool v = true;
   // Iterate over each iod entry in order:
-  for(unsigned int idx = 0; idx < niods; ++idx)
+  for(size_t idx = 0; idx < niods; ++idx)
     {
     const IODEntry &iodentry = iod.GetIODEntry(idx);
     const char *ref = iodentry.GetRef();
